requirements.txt
flask
docxtpl
python-docx
